requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # For PostgreSQL on Render

# File handling dependencies
python-multipart==0.0.6
PyMuPDF==1.23.8  # For PDF text extraction
python-magic-bin==0.4.14  # Windows compatible

# Additional utilities
python-dotenv==1.0.0
aiofiles==23.2.0
aiohttp==3.9.1
psutil==5.9.6

# OCR and text processing
spacy==3.7.2
Pillow==10.1.0

# Production server
gunicorn==21.2.0